Re-org directory structure
The layout in the source tree now better matches the installed layout,
which makes development easier; when the app is run from the source tree,
the web server can find the files it serves.

Add $(realpath) to shell scripts so that absolute paths are used, and so
we don't have to repeat ../ in so many places.